+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 3.3.1.4
 
 - Add in some more error checking for project id's from slugs
+- Fix not being able to get some Curse modpack slugs